# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cafetería",
        "description": "API para gestión de cafetería",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/turnos": {
            "get": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Obtener lista de turnos",
                "operationId": "6adc8ab3972fa2e0a40a9ddec9b865b9",
                "responses": {
                    "200": {
                        "description": "Lista completa de turnos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Turno"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Crear un nuevo turno",
                "operationId": "99bc216a5722c88def2faab3dfae2a13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "hora_inicio",
                                    "hora_fin"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "hora_inicio": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00"
                                    },
                                    "hora_fin": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "16:00"
                                    },
                                    "notas": {
                                        "type": "string",
                                        "example": "Turno de mañana"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Turno creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Turno"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/turnos/{id}": {
            "get": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Obtener un turno por ID",
                "operationId": "a226444e58c216d77952b74a0a1dddd0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del turno",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del turno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Turno"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Turno no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Actualizar un turno existente",
                "operationId": "b09db997e2dac70bb0fbe205c2038b60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del turno",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "hora_inicio": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00"
                                    },
                                    "hora_fin": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "16:00"
                                    },
                                    "notas": {
                                        "type": "string",
                                        "example": "Turno de mañana"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Turno actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Turno"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Turno no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Eliminar un turno",
                "operationId": "90a5cffdb27234b86eb4ab2e09bd9c44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del turno",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Turno eliminado exitosamente"
                    },
                    "404": {
                        "description": "Turno no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Turno": {
                "required": [
                    "usuario_id",
                    "hora_inicio",
                    "hora_fin"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "usuario_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "hora_inicio": {
                        "type": "string",
                        "format": "time",
                        "example": "08:00"
                    },
                    "hora_fin": {
                        "type": "string",
                        "format": "time",
                        "example": "16:00"
                    },
                    "notas": {
                        "type": "string",
                        "example": "Turno de mañana"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}